Rename frontend/src/utils/helpers.js -> frontend/src/helpers/chatAndMessageHelpers.js
