+++ frontend/src/helpers/chatAndMessageHelpers.js
